refactor: splash, onboarding, home을 features 폴더 내로 이동
