+++ lib/core/router.dart
@@ -16,16 +16,16 @@ import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/drea
 import 'package:mongbi_app/features/dream/presentation/pages/dream_write/dream_intro_page.dart';
 import 'package:mongbi_app/features/dream/presentation/pages/dream_write/dream_write_page.dart';
 import 'package:mongbi_app/features/history/presentation/pages/history_page.dart';
+import 'package:mongbi_app/features/home/home_page.dart';
+import 'package:mongbi_app/features/onboarding/onbording_exit_page.dart';
+import 'package:mongbi_app/features/onboarding/onbording_page.dart';
 import 'package:mongbi_app/features/setting/presentation/pages/alarm_setting_page.dart';
 import 'package:mongbi_app/features/setting/presentation/pages/nickname_input_page.dart';
 import 'package:mongbi_app/features/setting/presentation/pages/profile_setting_page.dart';
 import 'package:mongbi_app/features/setting/presentation/pages/setting_page.dart';
 import 'package:mongbi_app/features/setting/presentation/pages/widgets/open_source_license_page.dart';
+import 'package:mongbi_app/features/splash/splash_page.dart';
 import 'package:mongbi_app/features/statistics/presentation/statistics_page.dart';
-import 'package:mongbi_app/presentation/home/home_page.dart';
-import 'package:mongbi_app/presentation/onboarding/onbording_exit_page.dart';
-import 'package:mongbi_app/presentation/onboarding/onbording_page.dart';
-import 'package:mongbi_app/presentation/splash/splash_page.dart';
 
 final FirebaseAnalytics firebaseAnalytics = FirebaseAnalytics.instance;
 final GoRouter router = GoRouter(
