+++ lib/features/alarm/presentation/pages/alarm/widgets/delete_modal.dart
@@ -3,9 +3,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
 import 'package:mongbi_app/features/alarm/providers/alarm_provider.dart';
-import 'package:mongbi_app/presentation/common/button_type.dart';
-import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
-import 'package:mongbi_app/presentation/common/ghost_button_widget.dart';
+import 'package:mongbi_app/features/common/button_type.dart';
+import 'package:mongbi_app/features/common/filled_button_widget.dart';
+import 'package:mongbi_app/features/common/ghost_button_widget.dart';
 
 class DeleteModal extends StatelessWidget {
   const DeleteModal({super.key});
