+++ lib/features/alarm/presentation/pages/remind/remind_time_setting_page.dart
@@ -4,9 +4,9 @@ import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/analytics/analytics_helper.dart';
 import 'package:mongbi_app/features/alarm/presentation/pages/remind/widgets/remind_time_setting_image_widget.dart';
 import 'package:mongbi_app/features/alarm/presentation/pages/remind/widgets/remind_time_setting_text_widget.dart';
-import 'package:mongbi_app/presentation/common/button_type.dart';
-import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
-import 'package:mongbi_app/presentation/onboarding/widgets/onbording_skip_button_widget.dart';
+import 'package:mongbi_app/features/common/button_type.dart';
+import 'package:mongbi_app/features/common/filled_button_widget.dart';
+import 'package:mongbi_app/features/onboarding/widgets/onbording_skip_button_widget.dart';
 
 class RemindTimeSettingPage extends StatelessWidget {
   const RemindTimeSettingPage({super.key});
