+++ lib/features/auth/presentation/pages/social_login_page.dart
@@ -11,8 +11,8 @@ import 'package:mongbi_app/features/auth/presentation/pages/widgets/last_login_s
 import 'package:mongbi_app/features/auth/presentation/pages/widgets/mongbi_image_widget.dart';
 import 'package:mongbi_app/features/auth/presentation/pages/widgets/naver_login_button_widget.dart';
 import 'package:mongbi_app/features/auth/presentation/pages/widgets/text_widgets.dart';
+import 'package:mongbi_app/features/common/custom_snack_bar.dart';
 import 'package:mongbi_app/features/terms/presentation/pages/widgets/terms_bottom_sheet_layout_widget.dart';
-import 'package:mongbi_app/presentation/common/custom_snack_bar.dart';
 import 'package:mongbi_app/providers/auth_provider.dart';
 import 'package:mongbi_app/providers/last_login_provider.dart';
