+++ lib/features/common/filled_button_widget.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:mongbi_app/core/font.dart';
-import 'package:mongbi_app/presentation/common/button_type.dart';
+import 'package:mongbi_app/features/common/button_type.dart';
 
-class GhostButtonWidget extends StatelessWidget {
-  const GhostButtonWidget({
+class FilledButtonWidget extends StatelessWidget {
+  const FilledButtonWidget({
     super.key,
     required this.type,
     required this.text,
@@ -21,12 +20,19 @@ class GhostButtonWidget extends StatelessWidget {
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(999),
         boxShadow: [
-          BoxShadow(
-            color: Color(0x191A181B),
-            blurRadius: 10,
-            offset: Offset(2, 2),
-            spreadRadius: 0,
-          ),
+          onPress != null
+              ? BoxShadow(
+                color: Color(0x331A181B),
+                blurRadius: 10,
+                offset: Offset(2, 2),
+                spreadRadius: 0,
+              )
+              : BoxShadow(
+                color: Color(0x191A181B),
+                blurRadius: 10,
+                offset: Offset(2, 2),
+                spreadRadius: 0,
+              ),
         ],
       ),
       child: ElevatedButton(
@@ -48,8 +54,8 @@ class GhostButtonWidget extends StatelessWidget {
           ),
           overlayColor: WidgetStatePropertyAll(
             type == ButtonType.primary
-                ? Color.fromRGBO(219, 190, 255, 1)
-                : Color.fromRGBO(183, 235, 229, 1),
+                ? Color.fromRGBO(77, 25, 140, 1)
+                : Color.fromRGBO(13, 105, 94, 1),
           ),
           backgroundColor: WidgetStateProperty.resolveWith((states) {
             switch (type) {
@@ -58,32 +64,22 @@ class GhostButtonWidget extends StatelessWidget {
                   return Color.fromRGBO(245, 244, 245, 0.96);
                 }
 
-                return Color.fromRGBO(244, 234, 255, 1);
+                return Color.fromRGBO(140, 46, 255, 1);
 
               case ButtonType.secondary:
                 if (states.contains(WidgetState.disabled)) {
                   return Color.fromRGBO(245, 244, 245, 0.96);
                 }
 
-                return Color.fromRGBO(232, 249, 247, 1);
+                return Color.fromRGBO(23, 191, 171, 1);
             }
           }),
           foregroundColor: WidgetStateProperty.resolveWith((states) {
-            switch (type) {
-              case ButtonType.primary:
-                if (states.contains(WidgetState.disabled)) {
-                  return Color.fromRGBO(214, 212, 216, 1);
-                }
-
-                return Color.fromRGBO(178, 115, 255, 1);
-
-              case ButtonType.secondary:
-                if (states.contains(WidgetState.disabled)) {
-                  return Color.fromRGBO(214, 212, 216, 1);
-                }
-
-                return Color.fromRGBO(23, 191, 171, 1);
+            if (states.contains(WidgetState.disabled)) {
+              return Color.fromRGBO(214, 212, 216, 1);
             }
+
+            return Colors.white;
           }),
         ),
         child: Text(text),
