+++ lib/features/common/floating_animation_widget.dart
