+++ lib/features/common/touch_scale_widget.dart
