+++ lib/features/dream/presentation/pages/dream_analysis/dream_analysis_result_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
-import 'package:mongbi_app/presentation/common/floating_animation_widget.dart';
+import 'package:mongbi_app/features/common/floating_animation_widget.dart';
 
 class DreamAnalysisResultPage extends StatefulWidget {
   const DreamAnalysisResultPage({super.key, required this.isFirst});
