+++ lib/features/dream/presentation/pages/dream_write/dream_write_page.dart
@@ -3,10 +3,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
+import 'package:mongbi_app/features/common/button_type.dart';
+import 'package:mongbi_app/features/common/filled_button_widget.dart';
 import 'package:mongbi_app/features/dream/presentation/pages/dream_write/widgets/dream_content_input.dart';
 import 'package:mongbi_app/features/dream/presentation/pages/dream_write/widgets/mood_selection_row.dart';
-import 'package:mongbi_app/presentation/common/button_type.dart';
-import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
 import 'package:mongbi_app/providers/dream_provider.dart';
 
 class DreamWritePage extends ConsumerStatefulWidget {
