+++ lib/features/home/models/home_state.dart
