+++ lib/features/home/widgets/speech_bubble.dart
