+++ lib/features/setting/presentation/pages/widgets/remove_accont_modal.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
-import 'package:mongbi_app/presentation/common/button_type.dart';
-import 'package:mongbi_app/presentation/common/custom_snack_bar.dart';
-import 'package:mongbi_app/presentation/common/filled_button_widget.dart';
-import 'package:mongbi_app/presentation/common/ghost_button_widget.dart';
+import 'package:mongbi_app/features/common/button_type.dart';
+import 'package:mongbi_app/features/common/custom_snack_bar.dart';
+import 'package:mongbi_app/features/common/filled_button_widget.dart';
+import 'package:mongbi_app/features/common/ghost_button_widget.dart';
 import 'package:mongbi_app/providers/auth_provider.dart';
 
 class RemoveAccontModal extends ConsumerWidget {
