+++ lib/features/splash/splash_page.dart
@@ -3,9 +3,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/analytics/analytics_helper.dart';
 import 'package:mongbi_app/core/font.dart';
+import 'package:mongbi_app/features/common/floating_animation_widget.dart';
 import 'package:mongbi_app/features/setting/providers/user_info_provider.dart';
-import 'package:mongbi_app/presentation/common/floating_animation_widget.dart';
-import 'package:mongbi_app/presentation/splash/view_models/splash_status.dart';
+import 'package:mongbi_app/features/splash/view_models/splash_status.dart';
 
 class SplashPage extends ConsumerStatefulWidget {
   const SplashPage({super.key});
