+++ lib/features/splash/view_models/splash_state.dart
@@ -1,5 +1,5 @@
 import 'package:mongbi_app/features/setting/data/dtos/user_dto.dart';
-import 'package:mongbi_app/presentation/splash/view_models/splash_status.dart';
+import 'package:mongbi_app/features/splash/view_models/splash_status.dart';
 
 class SplashState {
   const SplashState({this.status = SplashStatus.initial, this.userList});
